Refactored LevelBuilder. Added some fail checks and error handling. Updated graphics.

+++ src/main/java/model/customLevelModels/CustomLevelModel.java
@@ -6,8 +6,13 @@ import java.io.IOException;
 import java.io.Writer;
 
 public class CustomLevelModel {
+        Boolean playerExists = false;
+        Boolean targetExists = false;
+        Boolean planetExists = false;
+        public Boolean levelBuilt = false;
 
         Writer writer;
+
         JsonArrayBuilder planetArrayBuilder = Json.createArrayBuilder();
         JsonArrayBuilder targetArrayBuilder = Json.createArrayBuilder();
         JsonArrayBuilder playerArrayBuilder = Json.createArrayBuilder();
@@ -22,9 +27,15 @@ public class CustomLevelModel {
                         case "Hazy Clouds" -> imagePath = "src/main/resources/img/hazyClouds.gif";
                         case "Light Green" -> imagePath = "src/main/resources/img/lightGreen.gif";
                         case "PlaKanyet West" -> imagePath = "src/main/resources/img/kanye.jpg";
+                        // Target Objects
                         // Player Objects
-                        case "Black and White" -> imagePath = "src/main/resources/img/blackWhitePlayer2.gif";
-                        default -> imagePath = "kanye";
+                        case "Black/White" -> imagePath = "src/main/resources/img/blackWhitePlayer.gif";
+                        case "Gold" -> imagePath = "src/main/resources/img/goldPlayer.gif";
+                        case "Red" -> imagePath = "src/main/resources/img/redPlayer.gif";
+                        case "White/Orange" -> imagePath = "src/main/resources/img/whiteOrangePlayer.gif";
+                        default -> imagePath = "src/main/resources/img/kanye.jpg";
+
+
                 }
                 return imagePath;
         }
@@ -47,28 +58,39 @@ public class CustomLevelModel {
                 String typeString = jsonLevelObject.getString("type");
                 if(typeString.equals("Planet")){
                         planetArrayBuilder.add(jsonLevelObject);
+                        planetExists = true;
                 }else if(typeString.equals("Target")){
                         targetArrayBuilder.add(jsonLevelObject);
+                        targetExists = true;
                 }else{
                         playerArrayBuilder.add(jsonLevelObject);
+                        playerExists = true;
                 }
         }
-
         public void createJson(String levelName) throws IOException {
-                JsonArray planetArray = planetArrayBuilder.build();
-                JsonArray targetArray = targetArrayBuilder.build();
-                JsonArray playerArray = playerArrayBuilder.build();
+                if(targetExists && playerExists && planetExists) {
 
-                JsonObject jsonFile = Json.createObjectBuilder()
-                        .add("planets", planetArray)
-                        .add("targets", targetArray)
-                        .add("players", playerArray)
-                        .build();
+                        JsonArray planetArray = planetArrayBuilder.build();
+                        JsonArray targetArray = targetArrayBuilder.build();
+                        JsonArray playerArray = playerArrayBuilder.build();
 
-                FileWriter fileWriter = new FileWriter("src/main/resources/json/levels/" + levelName + ".json");
-                JsonWriter jWrite = Json.createWriter(fileWriter);
-                jWrite.writeObject(jsonFile);
-                jWrite.close();
+                        JsonObject jsonFile = Json.createObjectBuilder()
+                                .add("planets", planetArray)
+                                .add("targets", targetArray)
+                                .add("players", playerArray)
+                                .build();
+
+                        FileWriter fileWriter = new FileWriter("src/main/resources/json/levels/" + levelName + ".json");
+                        JsonWriter jWrite = Json.createWriter(fileWriter);
+                        jWrite.writeObject(jsonFile);
+                        jWrite.close();
+                        levelBuilt = true;
+                }else{
+                        levelBuilt = false;
+                }
+        }
+        public boolean levelBuilt(){
+                return levelBuilt;
         }
 }
